Panels and Team Sync
The Solid Teaching Panel Team used to be a project. In this pull the team members are included in the panel and vice versa to make the information coherent. Feel free to remove yourself from the panel will a pull request if you are no longer interested in working on this material.