editors to focus on specification
Content creation for documentation currently falls under the editorial process. The specification work is very time consuming and perhaps we could give the editors more focus by allocating some of the other tasks to a dedicated team which I have called the 'creators' under solid project.org website. 

There isn't a clear description of what the roadmap so perhaps we can remove that from the editorial process until there is greater clarity on the scope of the roadmap.