Create repo-overview.md
Seeing as what goes on where and who is involved is linked to the process I propose keeping this overview in the process repo.  

Note that I have proposed Solid Administrators are responsible for updating this file not the Solid Director and anyone can submit an issue to update it.