Proposal for advancing issues while keeping editors in loop
Based on the experience in the past milestone, I suggest adding a "best practice" suggestion to the process, in which people submitting proposals are encouraged to notify editors of the progress. The idea is that this will be helpful for editors to feel in the loop, as we lacked this in this round, and that editors will use these notifications to advance issues on the project board. I purposely did not mention the project board though, as editors can be free to organize the work, but I think these phases are fairly universal. 

I think this proposal matches the discussion we had on the gap call too.