This Week in Solid Process
This process change proposal defines how to write, edit and send the newsletter. As part of this pull request, I propose including [The Editors](https://github.com/solid/process/blob/master/editors.md) as authors at the bottom of the newsletter and in the right-hand column of the table on https://github.com/solid/solidproject.org/blob/staging/_posts/this-week-in-solid/next.md which is the basis for every edition too.