De-list inactive panels for now
When panels were first introduced there was a wild proliferation of panels. This resulted in some trouble keeping track of what was going on as mentioned in https://github.com/solid/process/issues/135 

In the process it states that "Panel repositories that are inactive for more than six months may be archived by Solid Administrators." 

With this pull request, the inactive panels have been removed. This does NOT mean that the topics are unimportant or that they should not be picked up at a later stage. It