Update to correspond only to Active Panels
Many panels have been inactive for at least 6 months and have therefore been archived according to the process "Panel repositories that are inactive for more than six months may be archived by Solid Administrators." 

This pull request removes the editorial assignments that do not have active panels at the moment. Of course, as soon as there is more activity it would be a good idea to reopen the editorial assignments so that it is clear which editors should review proposals from these panels.