Make the process more digestible for the reader
This pull request is to make the process easier to read without changing any of the content or meaning of the actual process. Why is this important? Making the process easier to read means that it is more accessible and therefore more likely that it is actually used as a tool to help us work together. 

The section above detailing 'what is Solid?' is taken word for word from the website copy on solidproject.org submitted and approved by @timbl. The definitions of WebID and Pod are taken directly from solidproject.org/faq. This is not a pull request to debate what is Solid, of the key definitions, it is simply a place to record the latest copy. If you are interested in discussing the actual content I would encourage you to submit a pull request to the homepage or FAQ of the website and we can align this process copy if there are any agreed changes there.  

The definitions of specification, standard, and protocol are taken from the Merriam-Webster dictionary so if there is a prefered dictionary this would be possible to use.  Suggestions on preferred dictionaries welcome. 

Solid App, Solid community, Solid ecosystem, and Solid Panels are newly created definitions based on the copy in the process or deductions of it so here input would be welcome via this pull request. The reason I included these terms is because they are used very often on the chat forums so it would be helpful to have a go-to definition.

The Solid Team is taken from solidproject.org/team. This is not a pull request to discuss if we should call it Solid Team or something else, that should happen on the website. As soon as there is a change on the website we can update it here too. The intention here is to have an accurate record of how to make decisions together as a team. The text on the Solid Team is designed to summarise what was being repeated with every role in the previous process which made reading it very repetitive and numbing. By stating it once at the top the hope is that it will be more easily digestible by the reader.  

The explicit role definitions and listed responsibilities are an attempt to record the actual reality, feedback on this is welcome, also additional roles and responsibilities that I have not mentioned but you feel should be added.  

Under the Solid Specification, "stakeholders" and "contributors" have been moved up to definitions. 

Apart from isolating the role and responsibility descriptions along with how individuals can apply, get appointed and what they get access to, the copy around specification proposals and review of proposals is untouched.

Just to restate - the aim of this pull request is not to fundamentally change any of the processes but to make it easier to read so that more people can use it to work together effectively.