Minimised Version of #211
This pull request is another version of #211 where no text from other places, for example, solidproject.org or the Meriam Webster dictionary is included which was flagged as too much of a change to the existing process by @justinwb and @timbl. 

To reiterate - the aim of this pull request is to make the process easier to read so that it is more likely that we, as a team, clearly understand the agreement on how to work together. 

The ease of understanding is also relevant to the Solid community at large. For example, Aschijver explicitly stated on the forum that the process was too much of a burden to read and understand. The way that the process is currently written is asking too much too soon from contributors.