Lists not in sync
The lists are reflected in the teams on https://github.com/orgs/solid/teams/panels 

However, users are being added to teams without being added to the lists and users are added to the lists without being added to the teams. Keeping the lists and the teams in sync is proving difficult in practice.

Therefore I propose including a link to the teams rather than duplicating the list in this .md file.