Final review from Sarven change
For some reason, it will not let me commit the suggestions Sarven made, so I made it manually. I also saw the comment about the DEIT, so there were a few parts that said DEI Team. I updated them to DEIT. I believe that was the final comment to be changed for this PR to be merged.